Added Authentication Activity and modified Main Activity
