+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Daylee">
 
-        <activity android:name=".Dashboard" />
+        <activity android:name=".AuthenticationActivity"
+            android:theme="@style/Theme.Daylee.NoActionBar"/>
 
-        <activity android:name=".Splash_Screen"
+        <activity android:name=".Dashboard" />
+        
+        <activity
+            android:name=".Splash_Screen"
             android:theme="@style/Theme.Daylee.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -23,12 +27,10 @@
         <activity
             android:name=".RegisterActivity"
             android:theme="@style/Theme.Daylee.NoActionBar" />
-
         <activity
-            android:name=".MainActivity"
-            android:theme="@style/Theme.Daylee.NoActionBar" />
+            android:name=".MainActivity" />
 
-        <activity android:name=".AddNote"/>
+        <activity android:name=".AddNote" />
     </application>
 
 </manifest>
