+++ app/src/main/java/com/example/daylee/AuthenticationActivity.kt
@@ -1,0 +1,26 @@
+package com.example.daylee
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.Button
+
+class AuthenticationActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_authentication)
+
+        val btnRegLogin = findViewById(R.id.btnRegLogin) as Button
+
+        btnRegLogin.setOnClickListener {
+            startActivity(Intent(this,RegisterActivity::class.java))
+            overridePendingTransition(R.anim.slide_from_right,R.anim.slide_to_left)
+        }
+
+        val btnLogin = findViewById<Button>(R.id.btnLogin)
+        btnLogin.setOnClickListener {
+            startActivity(Intent(this,Dashboard::class.java))
+            overridePendingTransition(R.anim.slide_from_right,R.anim.slide_to_left)
+        }
+    }
+}
