+++ app/src/main/java/com/example/daylee/MainActivity.kt
@@ -10,18 +10,6 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val btnRegLogin = findViewById(R.id.btnRegLogin) as Button
-
-        btnRegLogin.setOnClickListener {
-            startActivity(Intent(this,RegisterActivity::class.java))
-            overridePendingTransition(R.anim.slide_from_right,R.anim.slide_to_left)
-        }
-
-        val btnLogin = findViewById<Button>(R.id.btnLogin)
-        btnLogin.setOnClickListener {
-            startActivity(Intent(this,Dashboard::class.java))
-            overridePendingTransition(R.anim.slide_from_right,R.anim.slide_to_left)
-        }
 
 
     }
